# backend/src/openapi/spec.yaml
openapi: 3.1.0
info:
  title: Sage Flashcards API
  description: |
    API for the Sage flashcard application with spaced repetition learning.
    Supports deck management, card creation, study sessions with FSRS algorithm,
    and AI-powered card generation.
  version: 1.0.0
  contact:
    email: support@sage.app
servers:
  - url: https://api.sage.app
    description: Production server
  - url: http://localhost:3001
    description: Development server
paths:
...
  /api/decks:
    get:
      operationId: listDecks
      summary: List user's flashcard decks
      description: Returns all decks belonging to the authenticated user with study statistics
      responses:
        '200':
          description: List of decks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeckListResponse'

    post:
      operationId: createDeck
      summary: Create a new flashcard deck
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
              properties:
                title:
                  type: string
                  description: Deck title
                  example: Spanish Vocabulary
                description:
                  type: string
                  description: Deck description
                isPublic:
                  type: boolean
                  default: false
                category:
                  type: string
                  example: Languages
                tags:
                  type: array
                  items:
                    type: string
      responses:
        '201':
          description: Deck created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeckResponse'
  /api/decks/{deckId}:
    get:
      operationId: getDeck
      summary: Get deck with all cards
      parameters:
        - name: deckId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deck details with cards
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeckWithCardsResponse'
...
  /api/public/decks:
    get:
      operationId: browsePublicDecks
      summary: Browse public decks
      description: Search and browse community-shared flashcard decks
      parameters:
        - name: search
          in: query
          schema:
            type: string
          description: Search term
        - name: category
          in: query
          schema:
            type: string
          description: Filter by category
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Public decks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicDecksResponse'
  /api/public/decks/{deckId}/clone:
    post:
      operationId: cloneDeck
      summary: Import a public deck to your library
      parameters:
        - name: deckId
          in: path
          required: true
          schema:
            type: string
      responses:
        '201':
          description: Deck cloned to user's library
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeckResponse'
components:
  schemas:
    Deck:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        cardCount:
          type: integer
        masteredCount:
          type: integer
        learningCount:
          type: integer
        newCount:
          type: integer
        dueCount:
          type: integer
        isPublic:
          type: boolean
        category:
          type: string
        tags:
          type: array
          items:
            type: string
    Card:
      type: object
      properties:
        id:
          type: string
        front:
          type: string
        back:
          type: string
        cardType:
          type: string
          enum: [flashcard, multiple_choice]
        options:
          type: array
          items:
            type: string
        state:
          type: string
          enum: [new, learning, review, relearning]
    DeckListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/Deck'
    DeckResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/Deck'
    DeckWithCardsResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          allOf:
            - $ref: '#/components/schemas/Deck'
            - type: object
              properties:
                cards:
                  type: array
                  items:
                    $ref: '#/components/schemas/Card'
...
    PublicDecksResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            decks:
              type: array
              items:
                allOf:
                  - $ref: '#/components/schemas/Deck'
                  - type: object
                    properties:
                      authorName:
                        type: string
                      averageRating:
                        type: number
                      downloadCount:
                        type: integer
            total:
              type: integer
            hasMore:
              type: boolean